refactors code so that tokenize() happens inside the Indexer class
