+++ crawler.py
@@ -12,9 +12,10 @@ class Crawler:
         self.__visited = []
         self.__internal_url_structure = {}
         self.__tokens = {}
-        self.__crawl()
+        self.__soups = self.__crawl()
 
     def __crawl(self):
+        soups = {}
         for url in self.__seed_urls:
             self.__frontier.append(self.__base_url + url)
             while len(self.__frontier) > 0:
@@ -26,7 +27,7 @@ class Crawler:
                 key = self.__find_doc_title(current_url, '/', '.')
                 self.__internal_url_structure[key] = []
 
-                self.__tokens[key] = self.__tokenize(soup)
+                soups[key] = soup
 
                 for internal_url in soup.find_all('a'):
                     current_internal_url = self.__base_url + internal_url.get('href')
@@ -41,6 +42,7 @@ class Crawler:
 
         self.__print_tokens()
         self.__print_internal_url_structure()
+        return soups
 
     def __sort_keys_in_dict(self, dict):
         return OrderedDict(sorted(dict.items()))
@@ -70,44 +72,5 @@ class Crawler:
     def get_internal_url_structure(self):
         return self.__internal_url_structure
 
-
-    def __tokenize(self, soup):
-        document_tokens = []
-
-        # removes all line breaks
-        strained_soup = self.__remove_tags(["br"], soup)
-
-        # only iterate over text content, should be obvious, but what everest
-        for text in strained_soup.body.find_all(text=True):
-
-            # ignore all links
-            if (text.parent.name != "a"):
-
-                # removes interpunction, in this case "." and ","
-                text = self.__remove_interpunction(str(text))
-
-                # splits into separate words
-                token_list = text.split()
-
-                # if list not empty
-                if token_list:
-                    token_list = self.__normalize_tokens(token_list)
-
-                    #concatenate all lists
-                    document_tokens.extend(token_list)
-
-        return document_tokens
-
-    def __remove_tags(self, unwanted_tags, soup):
-        for tag in unwanted_tags:
-            for match in soup.find_all(tag):
-                match.unwrap()
-        return soup
-
-    def __remove_interpunction(self, text):
-        text_neu = re.sub(r'[\,\.]', '', text)
-        return text_neu
-
-    def __normalize_tokens(self, token_list):
-        token_list = [element.lower() for element in token_list]
-        return token_list
+    def get_soups(self):
+        return self.__soups
